fix(table): resolve ESLint errors in state-persistence and action bar
- Extract UrlScope, UrlFiltersConfig, RawUrlTableState, UseUrlTableStateOptions,
  UseUrlTableStateResult into a dedicated types.ts to satisfy types-in-types-file rule
- Fix import sort order in use-table-state.ts, use-url-table-state.ts,
  share-table-url-button.tsx, and table-action-bar.tsx
- Replace || with ?? in table-action-bar condition (prefer-nullish-coalescing)
- Move module-scope test constants/helpers into nested describes to satisfy
  no-module-scope-test-setup rule (url-codecs, use-table-state, use-url-table-state,
  share-table-url-button tests)
- Fix unsafe any cast for vi.advanceTimersByTime.bind(vi)
