+++ javascript/packages/core/components/table/components/table-action-bar/table-action-bar.tsx
@@ -1,7 +1,7 @@
+import { ShareTableUrlButton } from '../share-table-url-button/share-table-url-button';
 import { ActiveFilterTagList } from './components/table-active-filter-tag-list/table-active-filter-tag-list';
 import { TableFilterMenu } from './components/table-filter-menu/table-filter-menu';
 import { TableSearchInput } from './components/table-search-input/table-search-input';
-import { ShareTableUrlButton } from '../share-table-url-button/share-table-url-button';
 import { ActionsContainer, Container, TrailingContentContainer } from './styled-components';
 
 import type { TableActionBarProps } from './types';
@@ -34,7 +34,7 @@ export function TableActionBar<T>({
 
         {configuration.middle}
 
-        {(configuration.trailing || (configuration.enableShareUrl && getShareUrl)) && (
+        {(configuration.trailing ?? (configuration.enableShareUrl && getShareUrl)) && (
           <TrailingContentContainer>
             {configuration.trailing}
             {configuration.enableShareUrl && getShareUrl && (
